plugins/gemini/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "getstream-plugins-gemini"
version = "0.1.0"
description = "Google Gemini plugin for GetStream"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "getstream[webrtc]>=2.3.0a6",
    "getstream-plugins-common>=0.1.1",
    "google-genai>=1.29.0",
    "numpy>=2.2.6",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
getstream = { workspace = true }
getstream-plugins-common = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.1",
]
